add child data class Rain, in Wind data class change speed type

+++ core/src/main/java/com/example/core/business/entities/CityWheatherChildEntities.kt
@@ -34,6 +34,6 @@ data class Weather(
 )
 
 data class Wind(
-        val speed: Int,
+        val speed: Float,
         val deg: Int
 )
